% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa.R
\name{browseHPA}
\alias{browseHPA}
\title{Browser the HPA page for a gene.}
\usage{
browseHPA(id)
}
\arguments{
\item{id}{A `character()` with one or multipel Ensembl gene
identifiers.}
}
\value{
Returns the URL(s) for the provided gene identifiers. If
    interactive, opens the URL(s).
}
\description{
Browser the HPA page for a gene.
}
\examples{

browseHPA("ENSG00000163435")
}
\author{
Laurent Gatto
}
