% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa.R
\name{getHpaDate}
\alias{getHpaDate}
\alias{getHpaVersion}
\alias{getHpaEnsembl}
\alias{allHparData}
\title{Obtain HPA data release information}
\usage{
getHpaDate()
}
\value{
A `character()` with the release information.
}
\description{
Obtain the data and release information of the Human Protein Atlas
(HPA) sets distributed by `hpar`. See
http://www.proteinatlas.org/about/releases for the HPA release
history.
}
\examples{

allHparData()

getHpaVersion()
getHpaDate()
getHpaEnsembl()
}
\author{
Laurent Gatto
}
