% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getHparOptions}
\alias{getHparOptions}
\title{Returns the \code{hpar} option.}
\usage{
getHparOptions()
}
\value{
A \code{list} with the \code{hpar} option(s).
}
\description{
Returns the \code{hapr} options. Currently, there is only
one, \code{hpadata}. See \code{\link{setHparOptions}} for
details and examples.
}
\author{
Laurent Gatto
}

