% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa.R
\name{getHpaDate}
\alias{getHpaDate}
\alias{getHpaVersion}
\alias{getHpaEnsembl}
\title{Obtain HPA release information}
\usage{
getHpaDate()
}
\value{
A `character()` with the release information.
}
\description{
Obtain release information about the Human Protein Atlas (HPA)
data, including version, release data and Ensembl version the data
is based on. Detailed release information is accessed on-line
using [utils::browseURL()]. Sew
http://www.proteinatlas.org/about/releases for the HPA release
history.
}
\examples{
getHpaVersion()
getHpaDate()
getHpaEnsembl()
}
\author{
Laurent Gatto
}
