\name{getHparOptions}
\alias{getHparOptions}
\title{Returns the \code{hpar} option.}
\usage{
  getHparOptions()
}
\value{
  A \code{list} with the \code{hpar} option(s).
}
\description{
  Returns the \code{hapr} options. Currently, there is only
  one, \code{hpadata}. See \code{\link{setHparOptions}} for
  details and examples.
}
\author{
  Laurent Gatto
}

