% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{setHparOptions}
\alias{setHparOptions}
\title{Sets the 'hpar' options}
\usage{
setHparOptions(hpadata = allHparData())
}
\arguments{
\item{hpadata}{A \code{character} defining the default data to be
queries. Choices are any of available \code{hpar} datasets,
available by calling \code{\link{allHparData}}, or any
unambiguous prefix.}
}
\value{
Returns the set value of the \code{hpar} option.  Used for
    its side effect.
}
\description{
Function used to set the \code{hpadata} \code{hpar} option.
}
\examples{
oldOpt <- getHparOptions()
oldOpt$hpadata
setHparOptions("hpaNormalTissue")
getHparOptions()
setHparOptions(oldOpt$hpadata)
}
\author{
Laurent Gatto
}

